MisionTres/ApiChatbot/requirements.txt
nltk
scikit-learn
deep-translator
python-docx

